move fake logic
