+++ go.mod
@@ -434,3 +434,5 @@ replace github.com/gogo/protobuf => github.com/regen-network/protobuf v1.3.3-alp
 replace github.com/fbsobreira/gotron-sdk => github.com/smartcontractkit/chainlink-tron/relayer/gotron-sdk v0.0.5-0.20251014143056-a0c6328c91e9
 
 replace github.com/smartcontractkit/chainlink-sui => github.com/smartcontractkit/chainlink-sui v0.0.0-20260527160341-aa3adc0abf67
+
+replace github.com/smartcontractkit/chainlink-solana/contracts => /Users/silaslenihan/Desktop/repos/chainlink-solana-develop/contracts
